+++ cmd/agnt/main.go
@@ -6,12 +6,13 @@ import (
 
 	"github.com/joho/godotenv"
 	"github.com/spf13/cobra"
+	"github.com/standardbeagle/agnt/internal/daemon"
 )
 
-const (
-	appName    = "agnt"
-	appVersion = "0.7.5"
-)
+const appName = "agnt"
+
+// appVersion can be overridden at build time with -ldflags="-X main.appVersion=x.y.z"
+var appVersion = "0.7.5"
 
 var rootCmd = &cobra.Command{
 	Use:   appName,
@@ -44,8 +45,8 @@ func init() {
 	rootCmd.AddCommand(daemonCmd)
 	rootCmd.AddCommand(sessionCmd)
 
-	// Version template
-	rootCmd.SetVersionTemplate(fmt.Sprintf("%s v%s\n", appName, appVersion))
+	// Custom version output that includes daemon version
+	rootCmd.SetVersionTemplate(getVersionString())
 }
 
 func main() {
@@ -63,4 +64,24 @@ func isTerminal(f *os.File) bool {
 		return false
 	}
 	return (stat.Mode() & os.ModeCharDevice) != 0
+}
+
+// getVersionString returns the version string including daemon version if available
+func getVersionString() string {
+	version := fmt.Sprintf("%s v%s\n", appName, appVersion)
+
+	// Try to get daemon version
+	client := daemon.NewClient()
+	if err := client.Connect(); err == nil {
+		defer client.Close()
+		if info, err := client.Info(); err == nil {
+			version += fmt.Sprintf("daemon v%s (uptime: %s)\n", info.Version, info.Uptime)
+		} else {
+			version += fmt.Sprintf("daemon: error getting info (%v)\n", err)
+		}
+	} else {
+		version += "daemon: not running\n"
+	}
+
+	return version
 }